Extract apollo resolvers to module
